moved pcdpass.spec.ts tests into devconnect.spec.ts
